Created the frontend of my login page
In Login.html I completed the front end of my login page which consists of a welcome message "Welcome to Penny-wise", the logo of my PWA which is in logo.png and the login section which requires a username and password. The whole login page has been styled in style.css, replicating my storyboard. A specific username (Software) and password (Engineering) has been assigned using Javascript in script,js, and if the user places incorrect details or empty spaces warning messages are sent such as "Incorrect username or password" and "Please enter login details".