Attempted to make my page that displays all budget logs made.
Within entries.html I added code that creates the layout based on my storyboard and the area that will display all the logs when the information is submitted.

In Penny-wise.html I edited the name of my past entries file and title as they were both had spelling errors. Additionally, I added code to the section where users can add budgets so it is styled. I also added code to the submit button so that it is linked to the past entries page when clicked after the form is completed.

Started coding within script2.js so that information that has been logs can transfer into the past entries page.

Script3.js is connected to entries.html and consists of the dynamics needed for the budget log to be displayed on the page. It holds the code on the delete button and other components that are to be displayed on the past entries page.

Lastly, the style3.css file consists of the styling for my past entries page so it replicates my storyboard. It has the same code for the title and navigation bar but has code added for the boxes that display the budget logs.