Continued working on my frontend.
Within entries.html I edited the layout to ensure it was compatible with my JavaScript and storyboard, I fixed the JavaScript file so that Penny-wsie.html and entries.html have the same JavaScript file, I also added a div class for where the budget logs will appear.

In Login.html I added a title to the page, naming it "Login".

Within Penny-wise.html I made various edits due to the JavaScript that was to be added so the budget logs are added in the entries.html file, this includes the functions so they correlate with the functions in the JavaScript. I fixed the submit button to add the function for the JavaScript and removed the window reference as it was no longer needed. Additionally, I turned the input points into a form so it will work with the JavaScript when putting the budget logs in the past entries page.

Script2.js was where I added the JavaScript for the budget logs, it ensures that the user inserts a date, and puts the budget log in a layout when it is placed on the new page.

I deleted script3.js as the code that is needed for entries.html is within script2.js along with the javascript needed for Penny-wise.html.

Within style2.css I styled all my expense input areas and changed the positions and sizes for some of the components to ensure a neat layout.

style3.css consists the styling for the budget logs however they were deleted as they are no longer needed due to the changes made in the JavaScript.