Attempted to add an edit button to the logs.
Within script2.js I added code that will create an edit button, allow users to edit the date, category and expense when the expense log has been submitted. This code allows for the button to be made, what sections are able to be edited and a save and cancel button after the log has been edited to save the changes or cancel them if the user no longer needs to edit. However, the button is not fully functional as it only allows for one log to be edited rather than all of them.

After deleting the budgeting components I edited the positioning of other elements including the total spending section in style.css. I was originally going to add another section for users to add a category and expense however that didn't go through, evident is the changes made to the CSS of that section.

Within style3.css I added styling for the  edit form after clicking the edit button.