Changed layout to better fit my server.js and attempted to start my database.
Within server.js I edited multiple parts in order to make sure that it correctly links with my main JavaScript file. This is unfortunately not fully functional yet.

I decided to place all my past entries on the same page as the form causing me to delete my entries.html p age and style3.css as they were no longer needed.

As I deleted the past entries page there was no need for a navigation page which was deleted from my Penny-wise.html file. Additionally I had to edit some of the variables and the structure of my page so it matches my JavaScript and made it easier for my to link it to my server.

Script2.css was heavily edited to match with my server.js, causing many of my variables to be changed around.

As I changed the whole layout of my PWA many alterations were made within style2.css. I deleted the four sections where users can create an expense log as well as the total spending section, meaning that many things were removed and deleted to ensure it was functional.