Got my backend to work and finished styling my page.
Changes have been displayed in expense.db as this contains my table which alters due to the adding, editing and deleting of logs.

query.sql was altered as I was writing queries to ensure that my database appears.

I made some edits in server.js as I discovered some errors that prevented it from working and made various other changes that helped in its functionality.

Within Penny-wise.html, I removed some of the unnecessary JavaScript references, which reduced congestion in the code and removed functions that were no longer needed.

To ensure that my JavaScript is compatible with my server I made some edits to items like variables to ensure it works seamlessly.

As I continued to edit the layout of my PWA multiple edits were made in style2.css including the styling for the expense logs and title.