Made some small edits to my JavaScript and CSS.
Within expense.db the changes are the adding and deleting of logs from the interface.

Within script2.js I deleted some of the comments placed as they were unnecessary.

Within style2.css I edited the sorting function so it is fixed on the page and doesn't move when scrolling the page, enhancing the usability of the components within the page.